5544 fix app crashing when creating stocktake

+++ src/database/DataTypes/StocktakeBatch.js
@@ -25,7 +25,7 @@ export class StocktakeBatch extends Realm.Object {
    * @param  {Realm}  database
    */
   destructor(database) {
-    if (this.snapshotNumberOfPacks === 0 && this.itemBatch.numberOfPacks === 0) {
+    if (this.snapshotNumberOfPacks === 0 && this.itemBatch && this.itemBatch.numberOfPacks === 0) {
       database.delete('ItemBatch', this.itemBatch);
     }
   }
@@ -75,6 +75,7 @@ export class StocktakeBatch extends Realm.Object {
    * @return  {boolean}
    */
   get isReducedBelowMinimum() {
+    if (!this.itemBatch) return false;
     const { totalQuantity: stockOnHand } = this.itemBatch;
 
     // Return true if |stockOnHand + stocktakebatch.difference| is negative.
@@ -92,6 +93,7 @@ export class StocktakeBatch extends Realm.Object {
    * @return  {boolean}
    */
   get isSnapshotOutdated() {
+    if (!this.itemBatch) return false;
     return this.snapshotTotalQuantity !== this.itemBatch.totalQuantity;
   }
